Fix v2 exisiting member signup

+++ src/main/java/com/USWCicrcleLink/server/global/config/SchedulerConfig.java
@@ -5,6 +5,9 @@ import com.USWCicrcleLink.server.email.domain.EmailToken;
 import com.USWCicrcleLink.server.email.repository.EmailTokenRepository;
 import com.USWCicrcleLink.server.profile.domain.Profile;
 import com.USWCicrcleLink.server.profile.repository.ProfileRepository;
+import com.USWCicrcleLink.server.user.domain.ExistingMember.ClubMemberTemp;
+import com.USWCicrcleLink.server.user.repository.ClubMemberTempRepository;
+import com.USWCicrcleLink.server.user.service.ClubMemberAccountStatusService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -24,19 +27,25 @@ public class SchedulerConfig {
     private final EmailTokenRepository emailTokenRepository;
     private final AplictRepository aplictRepository;
     private final ProfileRepository profileRepository;
+    private final ClubMemberTempRepository clubMemberTempRepository;
+    private final ClubMemberAccountStatusService clubMemberAccountStatusService;
 
 
     // 미인증 회원 삭제
     @Scheduled(cron = "0 0 * * * *") // 1시간 마다 실행
     @Transactional
     public void deleteExpiredTokens() {
-        long tokenCount = emailTokenRepository.count();
+
             LocalDateTime time = LocalDateTime.now().minusHours(1); // 만료시간 1시간 경과된 토큰 삭제
             List<EmailToken> tokens = emailTokenRepository.findAllByCertificationTimeBefore(time);
-            emailTokenRepository.deleteAll(tokens);
-            log.debug("만료된 이메일 토큰 삭제: 삭제된 이메일 토큰의 수 =  {}", tokens.size());
-    }
 
+            if(!tokens.isEmpty()){
+                emailTokenRepository.deleteAll(tokens);
+                log.debug("만료된 이메일 토큰 삭제");
+            }else{
+                log.debug("삭제할 이메일 토큰 존재하지않음");
+            }
+    }
 
     // 매일 00시(자정)에 실행
     // 최초 합 통보 후 4일이 지난 지원서 삭제
@@ -73,4 +82,26 @@ public class SchedulerConfig {
             log.debug("만료된 FCM 토큰이 없음");
         }
     }
+
+    // 매일 00시(자정)에 실행
+    // 7일이 지난 임시 회원과 관련된 데이터 삭제
+    @Scheduled(cron = "0 0 0 * * ?")
+    @Transactional
+    public void deleteExpiredClubMemberTemp() {
+
+        List<ClubMemberTemp> expiredClubMemberTempList = clubMemberTempRepository.findAllByClubMemberTempExpiryDateBefore(LocalDateTime.now());
+
+        if (!expiredClubMemberTempList.isEmpty()) {
+            log.debug("만료된 clubMemberTemp가  존재합니다. 관련된 accountStatus 조회 후 삭제를 시작합니다");
+            for (ClubMemberTemp expired : expiredClubMemberTempList) {
+                clubMemberAccountStatusService.deleteAccountStatus(expired);
+            }
+            clubMemberTempRepository.deleteAll(expiredClubMemberTempList);
+            log.debug("만료된 clubMemberTemp 전부 삭제 완료");
+        } else {
+            log.debug("삭제할 clubMemberTemp 없음");
+        }
+    }
+
+
 }
